microservices/user-service/requirements.txt
fastapi
pymongo
uvicorn
pydantic
fastapi-jwt-auth
python-multipart
aiocache 